+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking API changes
 
-- Do not fall back to blockchain adapter URL when fetching public configuration. (#628)
+- Do not fall back to blockchain adapter URL when fetching public configuration. (#628 #629)
 
 ### Dependency Upgrades