+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Refactor `RestTemplateConfig` to use `HttpClient 5` and improve proxy handling. (#626)
 - Replace deprecated `connect` with `connectAsync` in `StompClientService`. (#627)
+- Replace deprecated `SconeServicesProperties` and `GramineServicesProperties` to support the `teeFrameworkVersion` field. (#630)
 
 ### Breaking API changes